Move more methods into CommonQueryTrait
