+++ src/ORM/Query/UpdateQuery.php
@@ -18,6 +18,7 @@ namespace Cake\ORM\Query;
 
 use Cake\Database\Connection;
 use Cake\Database\Query\UpdateQuery as DbUpdateQuery;
+use Cake\Database\ValueBinder;
 use Cake\ORM\Table;
 
 class UpdateQuery extends DbUpdateQuery
@@ -34,7 +35,20 @@ class UpdateQuery extends DbUpdateQuery
     {
         parent::__construct($connection);
 
+        $this->setRepository($table);
         $this->addDefaultTypes($table);
-        $this->update($table->getTable());
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function sql(?ValueBinder $binder = null): string
+    {
+        if (empty($this->_parts['update'])) {
+            $repository = $this->getRepository();
+            $this->update($repository->getTable());
+        }
+
+        return parent::sql($binder);
     }
 }
